add models and websocket

+++ app/src/main/java/com/example/erudit/Modals/Player.java
@@ -1,7 +1,7 @@
 package com.example.erudit.Modals;
 
 public class Player {
-    long id;
+    int id;
     String name;
-    long score;
+    int score;
 }
